Pasar funciones a carpeta y archivo helper.
